Draft addition of code from vscode extension
This means that code is now duplicated.
The long term solution to this is to port the critical parts to
lean-client-js-browser or similar.
The files that should change between vscode-lean and lean-web-editor
are: extension.ts, index.ts. All other files should be exactly the same.

The next step is to get it to compile.